fix : build시 테스트 코드 오류 수정
